fixed circular deps
