+++ src/entities/CatalogNavigation/ui/Breadcrumbs/Breadcrumbs.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { AppLink } from "shared/ui/AppLink/AppLink"
 import { useSelector } from "react-redux"
 import { ChevronBack, MainHome } from "shared/assets/icons/others"
-import { RoutePath } from "shared/config/routeConfig/routeConfig"
+import { RoutePath } from "shared/config/routeConfig/const"
 import { classNames } from "shared/lib/classNames/classNames"
 import { getCurrentTree } from "../../model/selectors/sidebarNavigationSelectors"
 import { CurrentTreeItem, CurrentTreeItemType } from "../../model/types/catalogNavigationSchema"
