+++ src/entities/Product/ui/ProductCard/ProductCard.tsx
@@ -3,7 +3,7 @@ import { Label } from "shared/ui/Label/Label"
 import { Typography } from "shared/ui/Typography/Typography"
 import { classNames } from "shared/lib/classNames/classNames"
 import { useNavigate } from "react-router-dom"
-import { RoutePath } from "shared/config/routeConfig/routeConfig"
+import { RoutePath } from "shared/config/routeConfig/const"
 import productPlaceholder from "shared/assets/images/productPlaceholder.jpg"
 import styles from "./ProductCard.module.scss"
 import { Product } from "../../model/types"
