+++ src/pages/CategoryPage/ui/CategoryLinks/CategoryLinks.tsx
@@ -1,4 +1,4 @@
-import { RoutePath } from "shared/config/routeConfig/routeConfig"
+import { RoutePath } from "shared/config/routeConfig/const"
 import { classNames } from "shared/lib/classNames/classNames"
 import { AppLink } from "shared/ui/AppLink/AppLink"
 import { useNavigate } from "react-router-dom"
