--- src/shared/ui/Example Button/Button.testt.tsx
@@ -1,16 +1,0 @@
-/* import { render, screen } from "@testing-library/react"
-import { Button, ButtonTheme } from "shared/ui/Example Button/Button"
-
-describe("Button", () => {
-    test("Test render", () => {
-        render(<Button>TEST</Button>)
-        expect(screen.getByText("TEST")).toBeInTheDocument()
-    })
-
-    test("Test clear theme", () => {
-        render(<Button theme={ButtonTheme.CLEAR}>TEST</Button>)
-        expect(screen.getByText("TEST")).toHaveClass("clear")
-        screen.debug()
-    })
-})
- */
